Updated Social Security info in legal.md
Noticed that changing gender with SS office is as easy as ticking the correct-gendered box. For those completing legal name changes with SS, checking this box will save them the trouble of submitting an extra form later.
Included links to the pages on the SS website that guide you through this process (as well as their article on name change). Replaced the NCTE link which was broken.