Updated name and gender marker change info
Changed a lot, but I think it's easier to follow now.

Removed a lot of information that was out of date (ALEA rules about surgery letters, needing a lawyer to update birth certificate in every case, passport updates, etc). Also tried to make the page a little friendlier and simpler for those new to the process.